refactor epsilon greedy into own file
